+++ adapters/outgoing/scrapers/src/main/java/be/xplore/pricescraper/item/AhBeScraper.java
@@ -1,4 +1,4 @@
-package be.xplore.pricescraper.utils.scrapers;
+package be.xplore.pricescraper.item;
 
 import java.util.Optional;
 import org.jsoup.nodes.Document;
