+++ adapters/outgoing/scrapers/src/main/java/be/xplore/pricescraper/item/LocalDummyScraper.java
@@ -1,4 +1,4 @@
-package be.xplore.pricescraper.utils.scrapers;
+package be.xplore.pricescraper.item;
 
 
 import java.util.Optional;
