+++ adapters/outgoing/scrapers/src/test/java/scrapers/AhBeScraperTests.java
@@ -1,18 +1,20 @@
-package be.xplore.pricescraper.utils.scrapers;
+package scrapers;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import be.xplore.pricescraper.item.AhBeScraper;
+import be.xplore.pricescraper.item.Scraper;
 import java.io.IOException;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest(classes = AldiBeScraper.class)
-public class AldiBeScraperTests {
+@SpringBootTest(classes = AhBeScraper.class)
+public class AhBeScraperTests {
   private final Scraper scraper;
 
-  public AldiBeScraperTests(@Qualifier("scraper-aldi.be") Scraper scraper) {
+  public AhBeScraperTests(@Qualifier("scraper-ah.be") Scraper scraper) {
     this.scraper = scraper;
   }
 
@@ -23,7 +25,7 @@ public class AldiBeScraperTests {
 
   @Test
   void getItemResults() throws IOException {
-    var response = scraper.scrape("jupiler-15-st-3001592-1-0.article.html");
+    var response = scraper.scrape("wi445543/duvel-blond-4-pack");
     assertTrue(response.isPresent());
     assertNotNull(response.get());
     assertNotNull(response.get().title());
