+++ adapters/outgoing/scrapers/src/test/java/scrapers/AldiBeScraperTests.java
@@ -1,18 +1,20 @@
-package be.xplore.pricescraper.utils.scrapers;
+package scrapers;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import be.xplore.pricescraper.item.AldiBeScraper;
+import be.xplore.pricescraper.item.Scraper;
 import java.io.IOException;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest(classes = LidlBeScraper.class)
-public class LidlBeScraperTests {
+@SpringBootTest(classes = AldiBeScraper.class)
+public class AldiBeScraperTests {
   private final Scraper scraper;
 
-  public LidlBeScraperTests(@Qualifier("scraper-lidl.be") Scraper scraper) {
+  public AldiBeScraperTests(@Qualifier("scraper-aldi.be") Scraper scraper) {
     this.scraper = scraper;
   }
 
@@ -23,7 +25,7 @@ public class LidlBeScraperTests {
 
   @Test
   void getItemResults() throws IOException {
-    var response = scraper.scrape("bleekwater/p740192819");
+    var response = scraper.scrape("jupiler-15-st-3001592-1-0.article.html");
     assertTrue(response.isPresent());
     assertNotNull(response.get());
     assertNotNull(response.get().title());
