+++ adapters/outgoing/scrapers/src/test/java/scrapers/LocalDummyScraperTests.java
@@ -1,7 +1,9 @@
-package be.xplore.pricescraper.utils.scrapers;
+package scrapers;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
+import be.xplore.pricescraper.item.LocalDummyScraper;
+import be.xplore.pricescraper.item.Scraper;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.test.context.SpringBootTest;
