+++ core/src/main/java/be/xplore/pricescraper/scrapers/ItemScraper.java
@@ -1,0 +1,13 @@
+package be.xplore.pricescraper.scrapers;
+
+import be.xplore.pricescraper.dtos.ShopItem;
+import java.io.IOException;
+import java.util.Optional;
+
+public interface ItemScraper {
+  Optional<ShopItem> scrape(String itemIdentifier) throws IOException;
+
+  Optional<ShopItem> scrapeFromFullUrl(String url) throws IOException;
+
+  String getBaseUrl();
+}
